transfer models to assets directory
