+++ app/Console/Commands/UpdateCryptoPrice.php
@@ -2,11 +2,8 @@
 
 namespace App\Console\Commands;
 
-use App\Models\Crypto;
 use App\Services\Api\Command\CryptoUpdateMethod;
-use GuzzleHttp\Client;
 use Illuminate\Console\Command;
-use Mockery\Exception;
 
 class UpdateCryptoPrice extends Command
 {
