+++ app/Http/Controllers/API/Admin/AdminStockController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\API\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StockRequest;
 use App\Http\Resources\StockResource;
-use App\Models\Stock;
+use App\Models\Assets\Stock;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
 use Spatie\QueryBuilder\AllowedFilter;
