+++ app/Http/Controllers/Admin/Assets/AdminStockController.php
@@ -6,8 +6,8 @@ namespace App\Http\Controllers\Admin\Assets;
 use App\Exports\StocksExport;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StockRequest;
+use App\Models\Assets\Stock;
 use App\Models\Industry;
-use App\Models\Stock;
 use Illuminate\Database\Eloquent\Builder;
 use Maatwebsite\Excel\Facades\Excel;
 use Spatie\QueryBuilder\AllowedFilter;
