+++ app/Models/Assets/Crypto.php
@@ -1,28 +1,26 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Assets;
 
 use App\Enums\DirectionNameEnums;
+use App\Models\Direction;
 use App\Traits\HasDirection;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\SoftDeletes;
 
 /**
  * @property string $name
  * @property string $ticker
  * @property double $price
  * @property integer $lots
- * @property double $total_price
  */
-class Stock extends Model
+class Crypto extends Model
 {
+    use HasFactory, HasDirection;
 
-    use HasFactory;
-    use HasDirection;
-    use SoftDeletes;
-
-    protected $guarded = ['id'];
+    protected $guarded = [
+        'id'
+    ];
 
     public static function boot() {
         parent::boot();
@@ -30,14 +28,8 @@ class Stock extends Model
         static::creating(function ($model){
 
             $model->direction_id = Direction::query()
-                ->where('name', DirectionNameEnums::stocks()->value)
+                ->where('name', DirectionNameEnums::cryptos()->value)
                 ->first()?->id;
         });
     }
-
-    public function industry()
-    {
-        return $this->belongsTo(Industry::class);
-    }
-
 }
