+++ app/Models/Assets/Stock.php
@@ -1,8 +1,10 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Assets;
 
 use App\Enums\DirectionNameEnums;
+use App\Models\Direction;
+use App\Models\Industry;
 use App\Traits\HasDirection;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
@@ -12,14 +14,17 @@ use Illuminate\Database\Eloquent\SoftDeletes;
  * @property string $name
  * @property string $ticker
  * @property double $price
+ * @property integer $lots
+ * @property double $total_price
  */
-class Fund extends Model
+class Stock extends Model
 {
-    use HasFactory, HasDirection, SoftDeletes;
 
-    protected $guarded = [
-        'id'
-    ];
+    use HasFactory;
+    use HasDirection;
+    use SoftDeletes;
+
+    protected $guarded = ['id'];
 
     public static function boot() {
         parent::boot();
@@ -27,8 +32,14 @@ class Fund extends Model
         static::creating(function ($model){
 
             $model->direction_id = Direction::query()
-                ->where('name', DirectionNameEnums::funds()->value)
+                ->where('name', DirectionNameEnums::stocks()->value)
                 ->first()?->id;
         });
     }
+
+    public function industry()
+    {
+        return $this->belongsTo(Industry::class);
+    }
+
 }
