+++ app/Models/Industry.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Models\Assets\Stock;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
